isolate retry func in utils
